New database adaptations

+++ app/cronjob.js
@@ -17,28 +17,30 @@ const activatedCCS = new Cron.CronJob(
 	'00 00 10 * * 1-5', async () => { // At 10h from monday through friday 00 00 10 * * 1-5
 		const notifications = await db.getActivatedNotification();
 
-		if (notifications.length !== 0) { // checking if there is any notification to send
-			let currentCCS = { // loading data from first ccs
-				cod_ccs: notifications[0].conselho_id,
-				nome: await db.getNamefromCCS(notifications[0].conselho_id),
-				bairros: await db.getEveryBairro(notifications[0].conselho_id),
-			};
+		if (notifications) { // if there was any result
+			if (notifications.length !== 0) { // checking if there is any notification to send
+				let currentCCS = { // loading data from first ccs
+					cod_ccs: notifications[0].conselho_id,
+					nome: await db.getNamefromCCS(notifications[0].conselho_id),
+					bairros: await db.getEveryBairro(notifications[0].conselho_id),
+				};
 
 		for (const element of notifications) { // eslint-disable-line
-				if (element.conselho_id !== currentCCS.cod_ccs) { // check if we are not on the same CCS as before
-				// If we are not warning on the same CCS as before we have to reload the data
-				// This is an assurance in case more than one ccs gets activated
-				// Obs: the getActivatedNotification query orders results by the conselho_id
-					currentCCS = { // loading data from the new ccs
-						cod_ccs: element.conselho_id,
-						nome: await db.getNamefromCCS(element.conselho_id),
-						bairros: await db.getEveryBairro(element.conselho_id),
-					};
-				}
+					if (element.conselho_id !== currentCCS.cod_ccs) { // check if we are not on the same CCS as before
+						// If we are not warning on the same CCS as before we have to reload the data
+						// This is an assurance in case more than one ccs gets activated
+						// Obs: the getActivatedNotification query orders results by the conselho_id
+						currentCCS = { // loading data from the new ccs
+							cod_ccs: element.conselho_id,
+							nome: await db.getNamefromCCS(element.conselho_id),
+							bairros: await db.getEveryBairro(element.conselho_id),
+						};
+					}
 
-				// finally we send the messages
-				if (await broadcast.sendActivatedNotification(element.user_id, currentCCS.nome, currentCCS.bairros) === true) {
-					db.updateNotification(element.id); // table boolean gets updated if the message was sent succesfully
+					// finally we send the messages
+					if (await broadcast.sendActivatedNotification(element.user_id, currentCCS.nome, currentCCS.bairros) === true) {
+						db.updateNotification(element.id); // table boolean gets updated if the message was sent succesfully
+					}
 				}
 			}
 		}
@@ -63,54 +65,55 @@ const agendaChange = new Cron.CronJob(
 		console.log(notifications);
 
 		const date = new Date();
-
-		if (notifications.length !== 0) { // checking if there is any notification to send
+		if (notifications) { // if there was any result
+			if (notifications && notifications.length !== 0) { // checking if there is any notification to send
 				for (const element of notifications) { // eslint-disable-line
-				if (date > element.new_datahora) { // checks if reunion already happened (data_hora is 'behind' current time) (date > new_datahora)
+					if (date > element.new_datahora) { // checks if reunion already happened (data_hora is 'behind' current time) (date > new_datahora)
 					// updates notificado to TRUE (There's no need to warn the user anymore)
-					// It doesn't matter if there was a change to agendas.status or not
-					db.updateAgendaNotification(element.id);
+					// It doesn't matter if there was a change to agendas.status_id or not
+						db.updateAgendaNotification(element.id);
 
-					// finding labelAgenda_id from name
-					const ourLabels = await client.getLabelList(); // get all labels we have
-					const theOneLabel = await ourLabels.data.find(x => x.name === `agenda${element.agendas_id}`); // find the one label with the name same (we need the id)
+						// finding labelAgenda_id from name
+						const ourLabels = await client.getLabelList(); // get all labels we have
+						const theOneLabel = await ourLabels.data.find(x => x.name === `agenda${element.agendas_id}`); // find the one label with the name same (we need the id)
 
-					if (theOneLabel) { // if we have that label (we should always have it) we delete it
-						await client.deleteLabel(theOneLabel.id);
-					}
-				} else if (element.status !== 4) { // checks if there was any change in agenda
-					let message = ''; // the message that will be sent to the user depending on the case
-					switch (element.status) {
-					case 1: // reunion was canceled
-						message = `A reunião do ${element.ccs} agendada para ${help.formatDate(element.old_datahora)} no ${element.old_endereco} foi cancelada. ` +
+						if (theOneLabel) { // if we have that label (we should always have it) we delete it
+							await client.deleteLabel(theOneLabel.id);
+						}
+					} else if (element.status_id !== 4) { // checks if there was any change in agenda
+						let message = ''; // the message that will be sent to the user depending on the case
+						switch (element.status_id) {
+						case 1: // reunion was canceled
+							message = `A reunião do ${element.ccs} agendada para ${help.formatDate(element.old_datahora)} no ${element.old_endereco} foi cancelada. ` +
 						'Ainda não há nova data, mas você será notificado quando houver.';
-						// adding new entry to the table notificacao_agenda because user will be informed when this reunion is rescheduled (Status agenda must be 2)
-						await db.addAgenda(element.user_id, element.agendas_id, element.old_endereco, element.old_datahora.toLocaleString());
-						break;
-					case 2: // reunion was canceled and changed
-						message = `Há uma nova data para a reunião do ${element.ccs} que foi cancelada. Atenção para a mudança:\n\n` +
+							// adding new entry to the table notificacao_agenda because user will be informed when this reunion is rescheduled (status_id agenda must be 2)
+							await db.addAgenda(element.user_id, element.agendas_id, element.old_endereco, element.old_datahora.toLocaleString());
+							break;
+						case 2: // reunion was canceled and changed
+							message = `Há uma nova data para a reunião do ${element.ccs} que foi cancelada. Atenção para a mudança:\n\n` +
 							`🗓️ *Nova Data*: ${help.formatDate(element.new_datahora)}\n` +
 							`🏠 *Novo Local*: ${element.new_endereco}`;
-						break;
-					case 3: // reunion was canceled and changed
-						message = `Alterado: A reunião do ${element.ccs} agendada para ${help.formatDate(element.old_datahora)} no ${element.old_endereco}, foi alterada. ` +
+							break;
+						case 3: // reunion was canceled and changed
+							message = `Alterado: A reunião do ${element.ccs} agendada para ${help.formatDate(element.old_datahora)} no ${element.old_endereco}, foi alterada. ` +
 						'Atenção para a mudança:\n\n' +
 						`🗓️ *Nova Data*: ${help.formatDate(element.new_datahora)}\n` +
 						`🏠 *Novo Local*: ${element.new_endereco}`;
-						break;
-					default:
-					// unknow status?
-						break;
-					}
-					console.log(message);
-
-					if (message !== '') { // check if this is a known 'case'
-						if (await broadcast.sendAgendaNotification(element.user_id, message) === true) {
-							db.updateAgendaNotification(element.id); // table boolean gets updated if the message was sent succesfully
+							break;
+						default:
+							// unknow status_id?
+							break;
 						}
-					}
+						console.log(message);
+
+						if (message !== '') { // check if this is a known 'case'
+							if (await broadcast.sendAgendaNotification(element.user_id, message) === true) {
+								db.updateAgendaNotification(element.id); // table boolean gets updated if the message was sent succesfully
+							}
+						}
 					// sending the messages to the user
-				} // else: if the reunion hasn't happened already and there was no change (yet?) to the agenda.status there's nothing to do
+					} // else: if the reunion hasn't happened already and there was no change (yet?) to the agenda.status_id there's nothing to do
+				}
 			}
 		}
 	}, (() => {
